requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
binance==0.3.5
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect==1.3.0
dateparser==1.2.1
dnspython==2.7.0
dotenv==0.9.9
dynaconf==3.2.10
exceptiongroup==1.2.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
finvizfinance==1.1.0
frozendict==2.4.6
frozenlist==1.6.0
gnews==0.4.1
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
googlenewsdecoder==0.1.7
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.54
langchain-google-genai==2.1.3
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.32
lxml==5.3.2
lxml_html_clean==0.4.2
multidict==6.4.3
multitasking==0.0.11
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9.1
selectolax==0.3.28
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.67.1
tradingview-scraper==0.4.8
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.55
zstandard==0.23.0
